Move window functions and base classes into own package
