+++ h2/src/main/org/h2/command/Parser.java
@@ -175,16 +175,16 @@ import org.h2.expression.Wildcard;
 import org.h2.expression.aggregate.AbstractAggregate;
 import org.h2.expression.aggregate.Aggregate;
 import org.h2.expression.aggregate.Aggregate.AggregateType;
-import org.h2.expression.aggregate.DataAnalysisOperation;
 import org.h2.expression.aggregate.JavaAggregate;
-import org.h2.expression.aggregate.Window;
-import org.h2.expression.aggregate.WindowFrame;
-import org.h2.expression.aggregate.WindowFrameBound;
-import org.h2.expression.aggregate.WindowFrameBoundType;
-import org.h2.expression.aggregate.WindowFrameExclusion;
-import org.h2.expression.aggregate.WindowFrameUnits;
-import org.h2.expression.aggregate.WindowFunction;
-import org.h2.expression.aggregate.WindowFunctionType;
+import org.h2.expression.analysis.DataAnalysisOperation;
+import org.h2.expression.analysis.Window;
+import org.h2.expression.analysis.WindowFrame;
+import org.h2.expression.analysis.WindowFrameBound;
+import org.h2.expression.analysis.WindowFrameBoundType;
+import org.h2.expression.analysis.WindowFrameExclusion;
+import org.h2.expression.analysis.WindowFrameUnits;
+import org.h2.expression.analysis.WindowFunction;
+import org.h2.expression.analysis.WindowFunctionType;
 import org.h2.index.Index;
 import org.h2.message.DbException;
 import org.h2.result.SortOrder;
