+++ h2/src/main/org/h2/expression/analysis/PartitionData.java
@@ -3,7 +3,7 @@
  * and the EPL 1.0 (http://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
-package org.h2.expression.aggregate;
+package org.h2.expression.analysis;
 
 import java.util.HashMap;
