+++ h2/src/main/org/h2/expression/analysis/package.html
@@ -9,6 +9,6 @@ Initial Developer: H2 Group
 Javadoc package documentation
 </title></head><body style="font: 9pt/130% Tahoma, Arial, Helvetica, sans-serif; font-weight: normal;"><p>
 
-Aggregate functions.
+Base classes for data analysis operations and implementations of window functions.
 
 </p></body></html>
